Plymouth: extract generator functions into generator/plymouth.go
Mirrors the existing pattern (init/plymouth.go for the init side,
generator/console.go for vconsole). Moves addPlymouthSupport,
detectPlymouthTheme, plymouthPkgConfig, parseThemeImageDir, fcMatch,
parseThemeFonts, extractFontFamily, and detectHostGPUModules out of
generator/generator.go into a new generator/plymouth.go. Call sites
and GPU validation block stay in generator.go. Pure refactor — no
behavioural changes.
